+++ core/shared/src/main/scala/zio/Task.scala
@@ -107,7 +107,7 @@ object Task {
   /**
    * @see See [[zio.ZIO.collectAllSuccessesParN]]
    */
-  final def collectAllSuccessesParN[A](n: Long)(as: Iterable[Task[A]]): Task[List[A]] =
+  final def collectAllSuccessesParN[A](n: Int)(as: Iterable[Task[A]]): Task[List[A]] =
     ZIO.collectAllSuccessesParN(n)(as)
 
   /**
@@ -125,7 +125,7 @@ object Task {
   /**
    * @see See [[zio.ZIO.collectAllWithParN]]
    */
-  final def collectAllWithParN[A, B](n: Long)(as: Iterable[Task[A]])(f: PartialFunction[A, B]): Task[List[B]] =
+  final def collectAllWithParN[A, B](n: Int)(as: Iterable[Task[A]])(f: PartialFunction[A, B]): Task[List[B]] =
     ZIO.collectAllWithParN(n)(as)(f)
 
   /**
